fix: cast loop index to i32 for SQLite get_text in UidList
values.get_text expects i32 but the enumerate() index is usize, which
fails the native build. Cast to i32 like the get_int calls in the same
loop.

+++ native/src/core/su/db.rs
@@ -52,7 +52,7 @@ impl SqlTable for UidList {
                 "uid" => uid = values.get_int(i as i32),
                 "locked" => locked = values.get_int(i as i32) != 0,
                 "package_name" => {
-                    let pkg = values.get_text(i);
+                    let pkg = values.get_text(i as i32);
                     package_name.push_str(pkg);
                 }
                 _ => {}
